+++ BorjesLIA/BorjesLIA/Models/Euro/EuroExchangeModel.cs
@@ -43,7 +43,7 @@ namespace BorjesLIA.Models.Euro
         public int PlacingOrder { get; set; }
 
         [Display(Name = "Typ")]
-        public decimal Type { get; set; }
+        public int Type { get; set; }
 
         [Display(Name = "Aktiv")]
         public bool Active { get; set; }
